requirements.txt
torch
torchvision
matplotlib
scikit-learn
Pillow
gdown